fix: fix city Id from long to int

+++ sinaleasy-back/src/main/java/com/sinalez/sinaleasy_back/dtos/CityRecordDTO.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record CityRecordDTO(
-    @NotNull long id,
+    @NotNull Integer id,
     @NotBlank String name,
     @NotBlank String state,
     @NotNull  Integer rating
